Add utils func to check if an obj is a record
and use it in the new corpus class
